+++ backend/src/main/java/com/sepay/backend/payment/mapper/PaymentMapper.java
@@ -3,5 +3,4 @@ package com.sepay.backend.payment.mapper;
 import com.sepay.backend.payment.dto.PasswordDTO;
 
 public interface PaymentMapper {
-    int getPassword(int userNo);
 }
