+++ backend/src/main/java/com/sepay/backend/payment/service/PaymentService.java
@@ -7,5 +7,4 @@ import java.io.IOException;
 
 public interface PaymentService {
     byte[] createQR(String url) throws WriterException, IOException;
-    boolean checkPassword(PasswordDTO passwordDTO);
 }
