+++ backend/src/main/java/com/sepay/backend/user/mapper/UserMapper.java
@@ -5,4 +5,6 @@ import com.sepay.backend.user.dto.UserDTO;
 
 public interface UserMapper {
     UserDTO selectUser(Integer userNo);
+
+    int getPassword(int userNo);
 }
