+++ backend/src/main/java/com/sepay/backend/user/service/UserService.java
@@ -1,7 +1,10 @@
 package com.sepay.backend.user.service;
 
+import com.sepay.backend.payment.dto.PasswordDTO;
 import com.sepay.backend.user.dto.UserDTO;
 
 public interface UserService {
     UserDTO getInfo(Integer userNo);
+
+    boolean checkPassword(PasswordDTO passwordDTO);
 }
