+++ backend/src/main/java/com/sepay/backend/user/service/UserServiceImpl.java
@@ -1,5 +1,6 @@
 package com.sepay.backend.user.service;
 
+import com.sepay.backend.payment.dto.PasswordDTO;
 import com.sepay.backend.user.dto.UserDTO;
 import com.sepay.backend.user.mapper.UserMapper;
 import lombok.RequiredArgsConstructor;
@@ -17,5 +18,13 @@ public class UserServiceImpl implements UserService{
     @Override
     public UserDTO getInfo(Integer userNo) {
         return mapper.selectUser(userNo);
+    }
+
+    @Override
+    public boolean checkPassword(PasswordDTO passwordDTO) {
+        int userPassword = mapper.getPassword(1);
+        System.out.println(passwordDTO.getPassword());
+        System.out.println("??"+userPassword);
+        return false;
     }
 }
