+++ backend/src/main/resources/mapper/PaymentMapper.xml
@@ -4,10 +4,5 @@
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
 <mapper namespace="com.sepay.backend.payment.mapper.PaymentMapper">
-    <select id="getPassword" parameterType="int" resultType="int">
-        select second_pwd
-        from user
-        where user_no = #{userNo}
-    </select>
 
 </mapper>
