+++ backend/src/main/resources/mapper/UserMapper.xml
@@ -10,4 +10,9 @@
         where user_no = #{userNo}
     </select>
 
+    <select id="getPassword" parameterType="int" resultType="int">
+        select second_pwd
+        from user
+        where user_no = #{userNo}
+    </select>
 </mapper>
